Add --repo-update flag
Some users might have an outdated CocoaPods Specs repo, and because our pod is quite new it might fail to install Chatkit dependency. By adding flag `--repo-update` we can be sure that this will not happen.
There is just one downside, it might take longer to integrate the dependency into the project.