Remove async methods from SynchronizedArray
Over the course of multiple changes, the synchronized array has now arrived at
the only reliable implementation - a wrapper around a shared datastructure
with all access serialised via a *non-concurrent* queue.

There is no point submitting tasks to this queue asynchronously and passing in
a completion handler to be invoked when the operation is complete - all we
achieve doing this is a proliferation of callbacks and, more importantly, the
moving of the completion handler invocations in to the critical section. We
actually lose what ever concurrency there may be in the callers.