Don't allow newly de-serialised entities to propagate
They should be tested against the existing store and the existing entity used if there is one (after being updated with new fields).

This is because the entities hold important state (including the subscriptions, in the case of rooms), something which will be rmeoved from them in future to save this kind of error from being possible.