Keep scrolling as messages are added in the demo
If the last message is visible when we begin to update the table view,
then we should scroll to the bottom of the view when the update is
finished.

For reasons I cannot comprehend, this only works if the scroll is *not*
animated. If animation is requested, then the *second to last* message
is the one we scroll to, accompanied by a big "bounce" animation.