Prevent attempts to both remove and update a row in a single batch
When loading older messages, we were running an delete and an update on
a row with the same index within the same batch, when we added a message
at i, removed the date header at i+1 and then tried to update the
grouping of i+2, which now has index i+1 because of the removal of the
date header which was there. This crashes the app.

We now split the update batches up in to different operations for
safety, and do the bulk insert, the removal of the date header and the
fixing of the grouping in separate operations. The animations all still
overlap in time for a smooth visual transition.

The bracketing willUpdate and didUpdate callbacks are also moved in to a
function which takes a closure to execute between them so that the start
and end can be observed at the lexical block level.