Simplify auth based on feedback
Remove the provider "stack" concept, which is unnecessarily generic

All providers will need to be wrapped in both caching and retrying, so
the SDK should just provide these.

Allow for rotating session credentials

By passing functions in to the HTTPSTokenProvider to retrieve the
headers or query parameters to be added, we allow these to change over
time. For maximum flexibility, the response to these is asynchronous.